% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/400_Y_x_BTRR.R
\name{BTR_Y_x}
\alias{BTR_Y_x}
\title{Generalized-Dimension Bayesian Tensor Response Regression with Gaussian Graphical Model}
\usage{
BTR_Y_x(input, n.iter, n.burn, ranks, hyperparameters = NULL,
  save_after = NULL, save_llik = TRUE)
}
\arguments{
\item{input}{A list with at least two arguments, *Y* and *x*.The last two dimensions in the array *Y* should represent the time and subject, respectively. The first *D* dimensions of *Y* will represent the tensor dimensions at each time for each subject. *x* should either be a *n.time* by *subject* matrix, or a vector of length *n.time*.}

\item{n.iter}{The number of iterations that the MCMC will complete in total}

\item{n.burn}{The number of MCMC iterations that will be discarded as a burn-in}

\item{ranks}{The number of ranks for the model. See Guhaniyogi and Spencer (2018) for the full details.}

\item{hyperparameters}{A data frame with one row and six variables, with variables a.lambda, b.lambda, a.tau, b.tau, a.sig, and b.sig for the prior specifications. Default values are 3, 3^(1/(2*D)), D - 1, ranks^((1 / D) - 1), 1, and -log(0.95), respectively.}

\item{save_after}{The number of iterations after which to have the intermediate results saved. This can be helpful for debugging. The default does not save anything befor the end of the MCMC.}

\item{save_llik}{Should the log-likelihood be saved? This defaults to `TRUE`.}
}
\value{
A list of length 9 containing posterior draws for $\beta_{j,r}$, $W$, $\lambda_{j,r}$, $\Phi$, $\sigma_y^2$, and $\alpha$. The log-likeihood and acceptance ratio for the Metropolis-Hastings step are also included.
}
\description{
Generalized-Dimension Bayesian Tensor Response Regression with Gaussian Graphical Model
}
\examples{
BTRR_data <- Y_x_BTRR_data(3, 20, 100, c(30,30,30), 1, 1)
BTRR_results <- BTR_Y_x(input = temp$dat, 10, 0, 1)

}
